feat(data): Switch "Cafeteria 11-14 Uhr" and "To-Go" variant
Previously, "Cafeteria 11-14 Uhr" would be the label for "nmtisch".
Unfortunately, at 24th November 2021 this was changed to be the label
of "heisstheke", while "nmtisch" is now called
"Cafeteria 11-14 Uhr To-Go". It appears more sensible to represent the
status quo in bundled data instead of providing infinite backwards
compatibility with the name resolution.