feat: Export matchCanteenByName() and matchLineByName()
This functionality can be used by ka-mensa-api to fill in details of
past plans, when those plans were collected with an outdated set of
canteen and line names. In those cases, some of the stored 'id' fields
will be null. They can be filled in by re-matching the stored names with
updated data. By exposing the internal match functions, we enable this
functionality.